+++ src/cpp/src/llm/greedy_decoding.cpp
