+++ src/cpp/src/llm/group_beam_searcher.cpp
