+++ src/cpp/src/llm/llm_pipeline_base.hpp
