+++ src/python/openvino_genai/__init__.py
@@ -11,7 +11,7 @@ from .__version__ import __version__
 if hasattr(os, "add_dll_directory"):
     os.add_dll_directory(os.path.dirname(__file__))
 
-from .py_llm_pipeline import (
+from .py_openvino_genai import (
     ContinuousBatchingPipeline,
     DecodedResults,
     EncodedResults,
